+++ backends/carp_webservices/lib/carp_domain/carp_domain.g.dart
@@ -14,19 +14,25 @@ DeploymentServiceRequest _$DeploymentServiceRequestFromJson(
 }
 
 Map<String, dynamic> _$DeploymentServiceRequestToJson(
-    DeploymentServiceRequest instance) {
-  final val = <String, dynamic>{};
+        DeploymentServiceRequest instance) =>
+    <String, dynamic>{
+      r'$type': instance.$type,
+      'studyDeploymentId': instance.studyDeploymentId,
+    };
 
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull(r'$type', instance.$type);
-  writeNotNull('studyDeploymentId', instance.studyDeploymentId);
-  return val;
+GetActiveParticipationInvitations _$GetActiveParticipationInvitationsFromJson(
+    Map<String, dynamic> json) {
+  return GetActiveParticipationInvitations(
+    json['accountId'] as String,
+  )..$type = json[r'$type'] as String;
 }
+
+Map<String, dynamic> _$GetActiveParticipationInvitationsToJson(
+        GetActiveParticipationInvitations instance) =>
+    <String, dynamic>{
+      r'$type': instance.$type,
+      'accountId': instance.accountId,
+    };
 
 GetStudyDeploymentStatus _$GetStudyDeploymentStatusFromJson(
     Map<String, dynamic> json) {
@@ -36,19 +42,11 @@ GetStudyDeploymentStatus _$GetStudyDeploymentStatusFromJson(
 }
 
 Map<String, dynamic> _$GetStudyDeploymentStatusToJson(
-    GetStudyDeploymentStatus instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull(r'$type', instance.$type);
-  writeNotNull('studyDeploymentId', instance.studyDeploymentId);
-  return val;
-}
+        GetStudyDeploymentStatus instance) =>
+    <String, dynamic>{
+      r'$type': instance.$type,
+      'studyDeploymentId': instance.studyDeploymentId,
+    };
 
 RegisterDevice _$RegisterDeviceFromJson(Map<String, dynamic> json) {
   return RegisterDevice(
@@ -61,21 +59,13 @@ RegisterDevice _$RegisterDeviceFromJson(Map<String, dynamic> json) {
   )..$type = json[r'$type'] as String;
 }
 
-Map<String, dynamic> _$RegisterDeviceToJson(RegisterDevice instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull(r'$type', instance.$type);
-  writeNotNull('studyDeploymentId', instance.studyDeploymentId);
-  writeNotNull('deviceRoleName', instance.deviceRoleName);
-  writeNotNull('registration', instance.registration);
-  return val;
-}
+Map<String, dynamic> _$RegisterDeviceToJson(RegisterDevice instance) =>
+    <String, dynamic>{
+      r'$type': instance.$type,
+      'studyDeploymentId': instance.studyDeploymentId,
+      'deviceRoleName': instance.deviceRoleName,
+      'registration': instance.registration,
+    };
 
 UnregisterDevice _$UnregisterDeviceFromJson(Map<String, dynamic> json) {
   return UnregisterDevice(
@@ -84,42 +74,26 @@ UnregisterDevice _$UnregisterDeviceFromJson(Map<String, dynamic> json) {
   )..$type = json[r'$type'] as String;
 }
 
-Map<String, dynamic> _$UnregisterDeviceToJson(UnregisterDevice instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull(r'$type', instance.$type);
-  writeNotNull('studyDeploymentId', instance.studyDeploymentId);
-  val['deviceRoleName'] = instance.deviceRoleName;
-  return val;
-}
+Map<String, dynamic> _$UnregisterDeviceToJson(UnregisterDevice instance) =>
+    <String, dynamic>{
+      r'$type': instance.$type,
+      'studyDeploymentId': instance.studyDeploymentId,
+      'deviceRoleName': instance.deviceRoleName,
+    };
 
 DeviceRegistration _$DeviceRegistrationFromJson(Map<String, dynamic> json) {
   return DeviceRegistration(
-    registrationCreationDate: json['registrationCreationDate'] as int,
-    deviceId: json['deviceId'] as String,
+    json['deviceId'] as String,
+    json['registrationCreationDate'] as int,
   )..$type = json[r'$type'] as String;
 }
 
-Map<String, dynamic> _$DeviceRegistrationToJson(DeviceRegistration instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull(r'$type', instance.$type);
-  writeNotNull('registrationCreationDate', instance.registrationCreationDate);
-  writeNotNull('deviceId', instance.deviceId);
-  return val;
-}
+Map<String, dynamic> _$DeviceRegistrationToJson(DeviceRegistration instance) =>
+    <String, dynamic>{
+      r'$type': instance.$type,
+      'registrationCreationDate': instance.registrationCreationDate,
+      'deviceId': instance.deviceId,
+    };
 
 GetDeviceDeploymentFor _$GetDeviceDeploymentForFromJson(
     Map<String, dynamic> json) {
@@ -130,34 +104,52 @@ GetDeviceDeploymentFor _$GetDeviceDeploymentForFromJson(
 }
 
 Map<String, dynamic> _$GetDeviceDeploymentForToJson(
-    GetDeviceDeploymentFor instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull(r'$type', instance.$type);
-  writeNotNull('studyDeploymentId', instance.studyDeploymentId);
-  val['masterDeviceRoleName'] = instance.masterDeviceRoleName;
-  return val;
-}
+        GetDeviceDeploymentFor instance) =>
+    <String, dynamic>{
+      r'$type': instance.$type,
+      'studyDeploymentId': instance.studyDeploymentId,
+      'masterDeviceRoleName': instance.masterDeviceRoleName,
+    };
 
 DeploymentSuccessful _$DeploymentSuccessfulFromJson(Map<String, dynamic> json) {
   return DeploymentSuccessful(
     json['studyDeploymentId'] as String,
     json['masterDeviceRoleName'] as String,
-  )
-    ..$type = json[r'$type'] as String
-    ..deviceDeploymentLastUpdateDate =
-        json['deviceDeploymentLastUpdateDate'] as int;
+    json['deviceDeploymentLastUpdateDate'] as int,
+  )..$type = json[r'$type'] as String;
 }
 
 Map<String, dynamic> _$DeploymentSuccessfulToJson(
-    DeploymentSuccessful instance) {
-  final val = <String, dynamic>{};
+        DeploymentSuccessful instance) =>
+    <String, dynamic>{
+      r'$type': instance.$type,
+      'studyDeploymentId': instance.studyDeploymentId,
+      'masterDeviceRoleName': instance.masterDeviceRoleName,
+      'deviceDeploymentLastUpdateDate': instance.deviceDeploymentLastUpdateDate,
+    };
+
+ActiveParticipationInvitation _$ActiveParticipationInvitationFromJson(
+    Map<String, dynamic> json) {
+  return ActiveParticipationInvitation()
+    ..$type = json[r'$type'] as String
+    ..participation = json['participation'] == null
+        ? null
+        : Participation.fromJson(json['participation'] as Map<String, dynamic>)
+    ..invitation = json['invitation'] == null
+        ? null
+        : StudyInvitation.fromJson(json['invitation'] as Map<String, dynamic>)
+    ..devices = (json['devices'] as List)
+        ?.map((e) => e == null
+            ? null
+            : DeviceInvitation.fromJson(e as Map<String, dynamic>))
+        ?.toList();
+}
+
+Map<String, dynamic> _$ActiveParticipationInvitationToJson(
+    ActiveParticipationInvitation instance) {
+  final val = <String, dynamic>{
+    r'$type': instance.$type,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -165,11 +157,82 @@ Map<String, dynamic> _$DeploymentSuccessfulToJson(
     }
   }
 
-  writeNotNull(r'$type', instance.$type);
+  writeNotNull('participation', instance.participation);
+  writeNotNull('invitation', instance.invitation);
+  writeNotNull('devices', instance.devices);
+  return val;
+}
+
+Participation _$ParticipationFromJson(Map<String, dynamic> json) {
+  return Participation()
+    ..$type = json[r'$type'] as String
+    ..studyDeploymentId = json['studyDeploymentId'] as String
+    ..id = json['id'] as String
+    ..isRegistered = json['isRegistered'] as bool;
+}
+
+Map<String, dynamic> _$ParticipationToJson(Participation instance) {
+  final val = <String, dynamic>{
+    r'$type': instance.$type,
+  };
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
   writeNotNull('studyDeploymentId', instance.studyDeploymentId);
-  val['masterDeviceRoleName'] = instance.masterDeviceRoleName;
-  val['deviceDeploymentLastUpdateDate'] =
-      instance.deviceDeploymentLastUpdateDate;
+  writeNotNull('id', instance.id);
+  writeNotNull('isRegistered', instance.isRegistered);
+  return val;
+}
+
+StudyInvitation _$StudyInvitationFromJson(Map<String, dynamic> json) {
+  return StudyInvitation()
+    ..$type = json[r'$type'] as String
+    ..name = json['name'] as String
+    ..description = json['description'] as String
+    ..applicationData = json['applicationData'] as String;
+}
+
+Map<String, dynamic> _$StudyInvitationToJson(StudyInvitation instance) {
+  final val = <String, dynamic>{
+    r'$type': instance.$type,
+  };
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('name', instance.name);
+  writeNotNull('description', instance.description);
+  writeNotNull('applicationData', instance.applicationData);
+  return val;
+}
+
+DeviceInvitation _$DeviceInvitationFromJson(Map<String, dynamic> json) {
+  return DeviceInvitation()
+    ..$type = json[r'$type'] as String
+    ..deviceRoleName = json['deviceRoleName'] as String
+    ..isRegistered = json['isRegistered'] as bool;
+}
+
+Map<String, dynamic> _$DeviceInvitationToJson(DeviceInvitation instance) {
+  final val = <String, dynamic>{
+    r'$type': instance.$type,
+  };
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('deviceRoleName', instance.deviceRoleName);
+  writeNotNull('isRegistered', instance.isRegistered);
   return val;
 }
 
@@ -216,7 +279,9 @@ MasterDeviceDeployment _$MasterDeviceDeploymentFromJson(
 
 Map<String, dynamic> _$MasterDeviceDeploymentToJson(
     MasterDeviceDeployment instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    r'$type': instance.$type,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -224,7 +289,6 @@ Map<String, dynamic> _$MasterDeviceDeploymentToJson(
     }
   }
 
-  writeNotNull(r'$type', instance.$type);
   writeNotNull('configuration', instance.configuration);
   writeNotNull('connectedDevices', instance.connectedDevices);
   writeNotNull(
@@ -247,7 +311,9 @@ TaskDescriptor _$TaskDescriptorFromJson(Map<String, dynamic> json) {
 }
 
 Map<String, dynamic> _$TaskDescriptorToJson(TaskDescriptor instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    r'$type': instance.$type,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -255,9 +321,36 @@ Map<String, dynamic> _$TaskDescriptorToJson(TaskDescriptor instance) {
     }
   }
 
-  writeNotNull(r'$type', instance.$type);
   writeNotNull('name', instance.name);
   writeNotNull('measures', instance.measures);
+  return val;
+}
+
+CustomProtocolTask _$CustomProtocolTaskFromJson(Map<String, dynamic> json) {
+  return CustomProtocolTask()
+    ..$type = json[r'$type'] as String
+    ..name = json['name'] as String
+    ..measures = (json['measures'] as List)
+        ?.map((e) =>
+            e == null ? null : Measure.fromJson(e as Map<String, dynamic>))
+        ?.toList()
+    ..studyProtocol = json['studyProtocol'] as String;
+}
+
+Map<String, dynamic> _$CustomProtocolTaskToJson(CustomProtocolTask instance) {
+  final val = <String, dynamic>{
+    r'$type': instance.$type,
+  };
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('name', instance.name);
+  writeNotNull('measures', instance.measures);
+  writeNotNull('studyProtocol', instance.studyProtocol);
   return val;
 }
 
@@ -269,7 +362,9 @@ TriggerDescriptor _$TriggerDescriptorFromJson(Map<String, dynamic> json) {
 }
 
 Map<String, dynamic> _$TriggerDescriptorToJson(TriggerDescriptor instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    r'$type': instance.$type,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -277,7 +372,6 @@ Map<String, dynamic> _$TriggerDescriptorToJson(TriggerDescriptor instance) {
     }
   }
 
-  writeNotNull(r'$type', instance.$type);
   writeNotNull('sourceDeviceRoleName', instance.sourceDeviceRoleName);
   writeNotNull('requiresMasterDevice', instance.requiresMasterDevice);
   return val;
@@ -296,7 +390,9 @@ TriggeredTask _$TriggeredTaskFromJson(Map<String, dynamic> json) {
 }
 
 Map<String, dynamic> _$TriggeredTaskToJson(TriggeredTask instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    r'$type': instance.$type,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -304,7 +400,6 @@ Map<String, dynamic> _$TriggeredTaskToJson(TriggeredTask instance) {
     }
   }
 
-  writeNotNull(r'$type', instance.$type);
   writeNotNull('task', instance.task);
   writeNotNull('targetDevice', instance.targetDevice);
   return val;
@@ -328,7 +423,9 @@ StudyDeploymentStatus _$StudyDeploymentStatusFromJson(
 
 Map<String, dynamic> _$StudyDeploymentStatusToJson(
     StudyDeploymentStatus instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    r'$type': instance.$type,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -336,7 +433,6 @@ Map<String, dynamic> _$StudyDeploymentStatusToJson(
     }
   }
 
-  writeNotNull(r'$type', instance.$type);
   writeNotNull('studyDeploymentId', instance.studyDeploymentId);
   writeNotNull('devicesStatus', instance.devicesStatus);
   writeNotNull('startTime', instance.startTime?.toIso8601String());
@@ -364,7 +460,9 @@ DeviceDeploymentStatus _$DeviceDeploymentStatusFromJson(
 
 Map<String, dynamic> _$DeviceDeploymentStatusToJson(
     DeviceDeploymentStatus instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    r'$type': instance.$type,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -372,7 +470,6 @@ Map<String, dynamic> _$DeviceDeploymentStatusToJson(
     }
   }
 
-  writeNotNull(r'$type', instance.$type);
   writeNotNull('studyDeploymentId', instance.studyDeploymentId);
   writeNotNull('device', instance.device);
   writeNotNull('requiresDeployment', instance.requiresDeployment);
@@ -394,7 +491,9 @@ DeviceDescriptor _$DeviceDescriptorFromJson(Map<String, dynamic> json) {
 }
 
 Map<String, dynamic> _$DeviceDescriptorToJson(DeviceDescriptor instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    r'$type': instance.$type,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -402,7 +501,6 @@ Map<String, dynamic> _$DeviceDescriptorToJson(DeviceDescriptor instance) {
     }
   }
 
-  writeNotNull(r'$type', instance.$type);
   writeNotNull('studyDeploymentId', instance.studyDeploymentId);
   writeNotNull('isMasterDevice', instance.isMasterDevice);
   writeNotNull('roleName', instance.roleName);