+++ backends/carp_webservices/lib/carp_service/deployment_reference.dart
@@ -37,62 +37,54 @@ void registerFromJsonFunctions() {
   FromJsonFactory.registerFromJsonFunction("dk.cachet.carp.protocols.domain.triggers.TriggeredTask", TriggeredTask.fromJsonFunction);
 }
 
-/// Provide a deployment endpoint reference to a CARP web service. Used to:
-/// - get a deployment configuration for a master device (typically a phone)
-/// - set deployment success status
-/// - query deployment status
-/// - register and unregister this device as part of a deployment
+/// Provide a deployment endpoint reference to a CARP web service.
+///
+/// According to CARP core, the protocol for using the
+/// [deployment sub-system](https://github.com/cph-cachet/carp.core-kotlin/blob/develop/docs/carp-deployment.md) is:
+///
+///   - [status()] - get the study deployment status of this deployment.
+///   - [registerDevice()] - register this device - and associated devices - in this deployment
+///   - [get()] - get the deployment for this master device
+///   - [success()] - report the deployment as successful
+///   - [unRegisterDevice()] - unregister this - or other - device if no longer used
 class DeploymentReference extends CarpReference {
   /// The CARP study deployment ID.
   String get studyDeploymentId => service.app.study.deploymentId;
 
   /// The device role name for this master device.
-  ///
-  /// Default name is `phone`.
+  /// Default name is `Master`.
   String masterDeviceRoleName = 'Master';
 
-  /// The role name of this device as registered at CARP.
+  /// The role name of this device as specified in the [deployment] protocol.
   String deviceRoleName;
 
-  /// The registration at the CARP server, once this is done.
-  DeviceRegistration registration;
+  /// A unique id for this device as registered at CARP.
+  String deviceId;
 
-  DeploymentReference._(CarpService service, [this.masterDeviceRoleName = 'Master']) : super._(service) {
-    this.masterDeviceRoleName ??= 'Master';
-  }
+  MasterDeviceDeployment _deployment;
 
-  /// The URL for the deployment endpoint for this [DeploymentReference].
-  ///
-  /// {{PROTOCOL}}://{{SERVER_HOST}}:{{SERVER_PORT}}/api/deployments
-  String get deploymentEndpointUri => "${service.app.uri.toString()}/api/deployments";
+  /// The deployment for this master device, once fetched from the server.
+  MasterDeviceDeployment get deployment => _deployment;
 
-  /// The URL for the deployment status endpoint for this [DeploymentReference].
-  ///
-  /// {{PROTOCOL}}://{{SERVER_HOST}}:{{SERVER_PORT}}/api/deployments?deploymentId={{DEPLOYMENT_ID}}
-  String get deploymentStatusEndpointUri => "$deploymentEndpointUri?deploymentId=$studyDeploymentId";
-
-  /// The URL for the device deployment endpoint for this [DeploymentReference].
-  ///
-  /// {{PROTOCOL}}://{{SERVER_HOST}}:{{SERVER_PORT}}/api/deployments/{{DEPLYOMENT_ID}}/devices?masterDeviceRoleName=phone
-  String get deviceDeploymentEndpointUri => "$deploymentEndpointUri/$studyDeploymentId/devices?masterDeviceRoleName=$masterDeviceRoleName";
-
-  /// The URL for the deployment RPC endpoint for this [DeploymentReference].
-  ///
-  /// {{PROTOCOL}}://{{SERVER_HOST}}:{{SERVER_PORT}}/api/deployments/all
-  String get deploymentRPCEndpointUri => "${service.app.uri.toString()}/api/deployments/all";
+  DeploymentReference._(CarpService service, [this.masterDeviceRoleName = 'Master']) : super._(service);
 
   /// A generic RPC request to the CARP Server.
   Future<Map<String, dynamic>> _rpc(DeploymentServiceRequest request) async {
     final restHeaders = await headers;
     final String body = _encode(request.toJson());
 
-    print('REQUEST:\n$body');
-
-    http.Response response = await httpr.post(Uri.encodeFull(deploymentRPCEndpointUri), headers: restHeaders, body: body);
+    print('REQUEST: ${service.deploymentRPCEndpointUri}\n$body');
+    http.Response response = await httpr.post(Uri.encodeFull(service.deploymentRPCEndpointUri), headers: restHeaders, body: body);
     int httpStatusCode = response.statusCode;
-    Map<String, dynamic> responseJson = json.decode(response.body);
+    String responseBody = response.body;
+    print('RESPONSE: $httpStatusCode\n$responseBody');
 
-    print('RESPONSE:\n$httpStatusCode\n${response.body}');
+    // check if this is a json list, and if so turn it into a json map with one item called 'items'
+    if (responseBody.startsWith('[')) {
+      responseBody = '{"items": $responseBody}';
+    }
+
+    Map<String, dynamic> responseJson = json.decode(responseBody);
 
     if (httpStatusCode == HttpStatus.ok) {
       return responseJson;
@@ -108,35 +100,43 @@ class DeploymentReference extends CarpReference {
   /// Get the deployment status for this [DeploymentReference].
   Future<StudyDeploymentStatus> status() async => StudyDeploymentStatus.fromJson(await _rpc(GetStudyDeploymentStatus(studyDeploymentId)));
 
-  /// Register this device for this deployment at the CARP server.
-  ///  * [deviceRoleName] - the role name of this device, e.g. `phone`.
-  ///  * [deviceId] - a unique ID for this device.
+  /// Register a device for this deployment at the CARP server.
+  ///  * [deviceRoleName] - the role name of the device, e.g. `phone`.
+  ///  * [deviceId] - a unique id for this device. If not specified, a unique id will be generated and used.
   ///
   /// Returns the updated study deployment status if the registration is successful.
   /// Throws a [CarpServiceException] if not.
-  Future<StudyDeploymentStatus> registerDevice(String deviceRoleName, String deviceId) async {
-    assert(deviceRoleName != null, 'deviceRoleName has to be specified when registering a device in CARP.');
-    assert(deviceId != null, 'deviceId has to be specified when registering a device in CARP.');
+  Future<StudyDeploymentStatus> registerDevice({
+    @required String deviceRoleName,
+    String deviceId,
+  }) async {
+    assert(deviceRoleName != null && deviceRoleName.length > 0, 'deviceRoleName has to be specified when registering a device in CARP.');
+    deviceId ??= Uuid().v4().toString();
     this.deviceRoleName = deviceRoleName;
-
-    print('device role name: ${this.deviceRoleName}, $deviceRoleName');
-    registration = DeviceRegistration(deviceId: deviceId);
-    return StudyDeploymentStatus.fromJson(await _rpc(RegisterDevice(studyDeploymentId, deviceRoleName, registration)));
+    this.deviceId = deviceId;
+    return StudyDeploymentStatus.fromJson(await _rpc(RegisterDevice(studyDeploymentId, deviceRoleName, DeviceRegistration(deviceId))));
   }
 
-  /// Unregister this device for this deployment at the CARP server.
+  /// Unregister a device for this deployment at the CARP server.
+  ///  * [deviceRoleName] - the role name of the device, e.g. `phone`.
+  ///
+  /// If [deviceRoleName] is not specified, then the name specified in when registering
+  /// this device in the [registerDevice] method is used.
   ///
   /// Returns the updated study deployment status if the registration is successful.
   /// Throws a [CarpServiceException] if not.
-  Future<StudyDeploymentStatus> unRegisterDevice() async => StudyDeploymentStatus.fromJson(await _rpc(UnregisterDevice(studyDeploymentId, deviceRoleName)));
+  Future<StudyDeploymentStatus> unRegisterDevice([String deviceRoleName]) async =>
+      StudyDeploymentStatus.fromJson(await _rpc(UnregisterDevice(studyDeploymentId, (deviceRoleName != null) ? deviceRoleName : this.deviceRoleName)));
 
   /// Get the deployment for this [DeploymentReference] for the specified [masterDeviceRoleName].
-  /// Default to `Master` if not specified.
+  /// If [masterDeviceRoleName] is not specified, the previously used name is used, which is `Master` as default.
   Future<MasterDeviceDeployment> get([String masterDeviceRoleName]) async {
     if (masterDeviceRoleName != null) this.masterDeviceRoleName = masterDeviceRoleName;
-    return MasterDeviceDeployment.fromJson(await _rpc(GetDeviceDeploymentFor(studyDeploymentId, this.masterDeviceRoleName)));
+    _deployment = MasterDeviceDeployment.fromJson(await _rpc(GetDeviceDeploymentFor(studyDeploymentId, this.masterDeviceRoleName)));
+    return _deployment;
   }
 
   /// Mark this deployment as a success on the server.
-  Future<StudyDeploymentStatus> success() async => StudyDeploymentStatus.fromJson(await _rpc(DeploymentSuccessful(studyDeploymentId, masterDeviceRoleName)));
+  Future<StudyDeploymentStatus> success() async =>
+      StudyDeploymentStatus.fromJson(await _rpc(DeploymentSuccessful(studyDeploymentId, masterDeviceRoleName, deployment.lastUpdateDate)));
 }