+++ backends/carp_webservices/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.5.0
 * **BREAKING**: Updated to CAMS v. 0.9.0 with new `$type` json serialization
-* feature: support for deployment endpoints
+* feature: support for deployment endpoints using the `DeploymentReference` class
 * fix: OAuth token in `CarpUser` is now serialized ([#96](https://github.com/cph-cachet/carp.sensing-flutter/issues/96))
+* fix: fixed an error in the refresh token method.
 
 ## 0.4.6
 * feature: better support for authentication with stored tokens ([#79](https://github.com/cph-cachet/carp.sensing-flutter/issues/79)).